docs: deprecation notice
PROXY_GATEWAY_URL block and car backends were migrated
to Rainbow, bifrost-gateway is no longer maintained and runs old boxo
version.

The intention of this change is to make the last release
that will be picked up by people running docker `latest` release,
allowing them to migrate at their own convenience.

Once we have the last release, we can archive the repo, to reduce
confusion.

If someone still needs this, they can fork, but most of existing users
will be happy with Rainbow as drop-in replacement.